+++ BaiTapLonWinForm/Services/Implementations/TeacherService.cs
@@ -19,7 +19,7 @@ namespace BaiTapLonWinForm.Services.Implementations
         {
             _teacherRepository = teacherRepository;
         }
-        public Teacher getAllTeacherByClassId(long classId)
+        public Teacher getAllTeacherByClassId(int classId)
         {
             return _teacherRepository.GetAllTeacherByClassId(classId);
         }
